requirements.txt
streamlit-1.50.0
flask-3.1.2
neo4j-5.28.2
neo4j_graphrag-1.10.0
litellm-1.77.7
fitz-0.0.1.dev2
pymupdf4llm-0.0.27
sentence_transformers-5.1.1